fix: migration and relationship model

+++ app/Http/Controllers/LikeController.php
@@ -5,17 +5,18 @@ namespace App\Http\Controllers;
 use App\Models\Emotion;
 use App\Models\Like;
 use App\Models\Post;
-use Illuminate\Http\Request;
+use GuzzleHttp\Psr7\Request;
 use Illuminate\Support\Facades\Auth;
 
 class LikeController extends Controller
 {
-    public function LikePost(Post $post , Emotion $emotion)
-    {
-        $user = Auth::user(); 
-        $existingLike = Like::where('user_id', $user->id)
-        ->where('post_id', $post->id)
-        ->where('emotion_id', $emotion->id)->first();
+    public function LikePost(Request $request, Post $post, $emotion){
+        $user = Auth::user();
+        $validEmotions = ['like', 'love', 'haha', 'wow', 'sad', 'angry'];
+        if (!in_array($emotion, $validEmotions)){
+            return response()->json(['error' => 'Invalid emotion type'], 400);
+        }
+        $existingLike = Like::where('user_id', $user->id)->where('post_id', $post->id)->where('emotion', $emotion)->first();
         if ($existingLike) {
             $existingLike->delete();
             $message = 'Unliked successfully';
@@ -23,7 +24,7 @@ class LikeController extends Controller
             Like::create([
                 'user_id' => $user->id,
                 'post_id' => $post->id,
-                'emotion_id' => $emotion->id,
+                'emotion' => $emotion,
             ]);
             $message = 'Liked successfully';
         }
